Locate multiple header groups

This function adds direction annations to tidyxl data frame, specifying annotions for several headers at once. The resulting data frame will contain the following columns: .value, .direction, and .header_label. These columns indicate the value, direction and name of the header columns that will be produced in the tidy data frame produced by migrate.

locate_groups(sheet = NULL, direction = "N",
  .groupings = groupings(fmt_alignment_indent),
  .hook_if = hook_if(any(FALSE)), .hook_if_rev = hook_if(any(FALSE)),
  default_col_header_direction = "N",
  default_row_header_direction = "W", header_fill = "local_format_id")

Arguments

sheet
data frame created by xlsx_cells

direction
a string indicating which type of headers are to be labelled. Options include compass direction or up/down/left/right.

.groupings
expressions representing how header cells are differentiated. Most naturally works with fmt_* functions. These expressions must be provided within groupings.

.hook_if
expression determining whether directions are hooked. That is, whether a header cell to the left of data cells should be annotated as "W" or "WNW". This argument contains an expression within the function hook. This expression must evaluate to a single value for each header group. For example, see below that any(fmt_alignment_indent == 0) is used rather that fmt_alignment_indent == 0.

.hook_if_rev
expression determining whether directions are hooked. That is, whether a header cell to the left of data cells should be annotated as "W" or "WNW". This argument contains an expression within the function hook. This expression must evaluate to a single value for each header group. For example, see below that any(fmt_alignment_indent == 0) is used rather that fmt_alignment_indent == 0.

default_col_header_direction
Indicates which direction is given to col headers by default. Only need if "NNW" is required, rather than "N".

default_row_header_direction
Indicates which direction is given to row headers by default. Only need if "WNW" is required, rather than "W".

header_fill
deals with merged cells. Fills in neighbouring cells if they have the same "local_format_id", "style" or are within "borders".

Examples